+++ src/main/java/net/oneandone/pommes/repository/Repository.java
@@ -55,6 +55,10 @@ public interface Repository {
         if (repository != null) {
             return repository;
         }
+        repository = InlineRepository.createOpt(environment.gson(), url);
+        if (repository != null) {
+            return repository;
+        }
         file = world.file(url);
         if (file.exists()) {
             return new NodeRepository(environment, file, log);